+++ bin/check-rspec.rb
@@ -82,11 +82,6 @@ class CheckRspec < Sensu::Plugin::Check::CLI
          long: '--proxy-client SOURCE',
          required: false
 
-  option :index_results,
-         description: 'Append extra index value to rspec results',
-         long: '--index-results',
-         default: false
-
   def sensu_client_socket(msg)
     u = UDPSocket.new
     u.send(msg + "\n", 0, '127.0.0.1', 3030)
@@ -110,12 +105,7 @@ class CheckRspec < Sensu::Plugin::Check::CLI
     parsed        = JSON.parse(rspec_results)
 
     parsed['examples'].each_with_index do |rspec_test, index|
-      test_name = case config[:index_results]
-                  when true
-                    rspec_test['file_path'].split('/')[-1] + '_' + rspec_test['line_number'].to_s + '_' + index.to_s
-                  else
-                    rspec_test['file_path'].split('/')[-1] + '_' + rspec_test['line_number'].to_s
-                  end
+      test_name = rspec_test['file_path'].split('/')[-1] + '_' + rspec_test['line_number'].to_s + '_' + index.to_s
       output = rspec_test['full_description']
 
       if rspec_test['status'] == 'passed'